add c api

+++ mindspore/lite/src/inner_kernel.h
@@ -168,14 +168,14 @@ class InnerKernel : public Kernel {
   void set_in_tensors(const std::vector<lite::Tensor *> &in_tensors) { this->in_tensors_ = in_tensors; }
 
   virtual void set_in_tensor(lite::Tensor *in_tensor, int index) {
-    MS_ASSERT(index < in_tensors_.size());
+    MS_ASSERT(static_cast<size_t>(index) < in_tensors_.size());
     this->in_tensors_[index] = in_tensor;
   }
 
   void set_out_tensors(const std::vector<lite::Tensor *> &out_tensors) { this->out_tensors_ = out_tensors; }
 
   virtual void set_out_tensor(lite::Tensor *out_tensor, int index) {
-    MS_ASSERT(index < out_tensors_.size());
+    MS_ASSERT(static_cast<size_t>(index) < out_tensors_.size());
     this->out_tensors_[index] = out_tensor;
   }
